Split out a bunch of the results into their own requisite class files
