+++ lib/cucumber/core/test/result.rb
@@ -3,6 +3,17 @@
 require 'cucumber/messages'
 require 'cucumber/messages/helpers/time_conversion'
 
+require_relative 'result/raisable'
+
+require_relative 'result/ambiguous'
+require_relative 'result/failed'
+require_relative 'result/flaky'
+require_relative 'result/passed'
+require_relative 'result/pending'
+require_relative 'result/skipped'
+require_relative 'result/undefined'
+require_relative 'result/unknown'
+
 module Cucumber
   module Core
     module Test
@@ -28,278 +39,6 @@ module Cucumber
                 possible_result_type == to_sym
               end
             end
-          end
-        end
-
-        # Null object for results. Represents the state where we haven't run anything yet
-        class Unknown
-          include Result.query_methods :unknown
-
-          def describe_to(_visitor, *_args)
-            self
-          end
-
-          def with_filtered_backtrace(_filter)
-            self
-          end
-
-          def to_message
-            Cucumber::Messages::TestStepResult.new(
-              status: Cucumber::Messages::TestStepResultStatus::UNKNOWN,
-              duration: UnknownDuration.new.to_message_duration
-            )
-          end
-        end
-
-        class Passed
-          include Result.query_methods :passed
-
-          attr_accessor :duration
-
-          def self.ok?(*)
-            true
-          end
-
-          def initialize(duration)
-            raise ArgumentError unless duration
-
-            @duration = duration
-          end
-
-          def describe_to(visitor, *)
-            visitor.passed(*)
-            visitor.duration(duration, *)
-            self
-          end
-
-          def to_s
-            '✓'
-          end
-
-          def to_message
-            Cucumber::Messages::TestStepResult.new(
-              status: Cucumber::Messages::TestStepResultStatus::PASSED,
-              duration: duration.to_message_duration
-            )
-          end
-
-          def ok?(*)
-            self.class.ok?
-          end
-
-          def with_appended_backtrace(_step)
-            self
-          end
-
-          def with_filtered_backtrace(_filter)
-            self
-          end
-        end
-
-        class Failed
-          include Result.query_methods :failed
-
-          attr_reader :duration, :exception
-
-          def self.ok?(*)
-            false
-          end
-
-          def initialize(duration, exception)
-            raise ArgumentError unless duration
-            raise ArgumentError unless exception
-
-            @duration = duration
-            @exception = exception
-          end
-
-          def describe_to(visitor, *)
-            visitor.failed(*)
-            visitor.duration(duration, *)
-            visitor.exception(exception, *) if exception
-            self
-          end
-
-          def to_s
-            '✗'
-          end
-
-          def to_message
-            begin
-              message = exception.backtrace.join("\n")
-            rescue NoMethodError
-              message = ''
-            end
-
-            Cucumber::Messages::TestStepResult.new(
-              status: Cucumber::Messages::TestStepResultStatus::FAILED,
-              duration: duration.to_message_duration,
-              message: message
-            )
-          end
-
-          def ok?(*)
-            self.class.ok?
-          end
-
-          def with_duration(new_duration)
-            self.class.new(new_duration, exception)
-          end
-
-          def with_appended_backtrace(step)
-            exception.backtrace << step.backtrace_line if step.respond_to?(:backtrace_line)
-            self
-          end
-
-          def with_filtered_backtrace(filter)
-            self.class.new(duration, filter.new(exception.dup).exception)
-          end
-        end
-
-        # Flaky is not used directly as an execution result, but is used as a
-        # reporting result type for test cases that fails and the passes on
-        # retry, therefore only the class method self.ok? is needed.
-        class Flaky
-          def self.ok?
-            false
-          end
-        end
-
-        # Base class for exceptions that can be raised in a step definition causing the step to have that result.
-        class Raisable < StandardError
-          attr_reader :message, :duration
-
-          def initialize(message = '', duration = UnknownDuration.new, backtrace = nil)
-            @message = message
-            @duration = duration
-            super(message)
-            set_backtrace(backtrace) if backtrace
-          end
-
-          def with_message(new_message)
-            self.class.new(new_message, duration, backtrace)
-          end
-
-          def with_duration(new_duration)
-            self.class.new(message, new_duration, backtrace)
-          end
-
-          def with_appended_backtrace(step)
-            return self unless step.respond_to?(:backtrace_line)
-
-            set_backtrace([]) unless backtrace
-            backtrace << step.backtrace_line
-            self
-          end
-
-          def with_filtered_backtrace(filter)
-            return self unless backtrace
-
-            filter.new(dup).exception
-          end
-
-          def ok?
-            self.class.ok?
-          end
-        end
-
-        class Ambiguous < Raisable
-          include Result.query_methods :ambiguous
-
-          def self.ok?(*)
-            false
-          end
-
-          def describe_to(visitor, *)
-            visitor.ambiguous(*)
-            visitor.duration(duration, *)
-            self
-          end
-
-          def to_s
-            'A'
-          end
-
-          def to_message
-            Cucumber::Messages::TestStepResult.new(
-              status: Cucumber::Messages::TestStepResultStatus::AMBIGUOUS,
-              duration: duration.to_message_duration
-            )
-          end
-        end
-
-        class Undefined < Raisable
-          include Result.query_methods :undefined
-
-          def self.ok?
-            false
-          end
-
-          def describe_to(visitor, *)
-            visitor.undefined(*)
-            visitor.duration(duration, *)
-            self
-          end
-
-          def to_s
-            '?'
-          end
-
-          def to_message
-            Cucumber::Messages::TestStepResult.new(
-              status: Cucumber::Messages::TestStepResultStatus::UNDEFINED,
-              duration: duration.to_message_duration
-            )
-          end
-        end
-
-        class Skipped < Raisable
-          include Result.query_methods :skipped
-
-          def self.ok?(*)
-            true
-          end
-
-          def describe_to(visitor, *)
-            visitor.skipped(*)
-            visitor.duration(duration, *)
-            self
-          end
-
-          def to_s
-            '-'
-          end
-
-          def to_message
-            Cucumber::Messages::TestStepResult.new(
-              status: Cucumber::Messages::TestStepResultStatus::SKIPPED,
-              duration: duration.to_message_duration
-            )
-          end
-        end
-
-        class Pending < Raisable
-          include Result.query_methods :pending
-
-          def self.ok?
-            false
-          end
-
-          def describe_to(visitor, *)
-            visitor.pending(self, *)
-            visitor.duration(duration, *)
-            self
-          end
-
-          def to_s
-            'P'
-          end
-
-          def to_message
-            Cucumber::Messages::TestStepResult.new(
-              status: Cucumber::Messages::TestStepResultStatus::PENDING,
-              duration: duration.to_message_duration
-            )
           end
         end
