+++ lib/cucumber/core/test/result/unknown.rb
@@ -1,0 +1,32 @@
+# frozen_string_literal: true
+
+require 'cucumber/messages'
+require 'cucumber/messages/helpers/time_conversion'
+
+module Cucumber
+  module Core
+    module Test
+      module Result
+        # Null object for results. Represents the state where we haven't run anything yet
+        class Unknown
+          include Result.query_methods :unknown
+
+          def describe_to(_visitor, *_args)
+            self
+          end
+
+          def with_filtered_backtrace(_filter)
+            self
+          end
+
+          def to_message
+            Cucumber::Messages::TestStepResult.new(
+              status: Cucumber::Messages::TestStepResultStatus::UNKNOWN,
+              duration: UnknownDuration.new.to_message_duration
+            )
+          end
+        end
+      end
+    end
+  end
+end
